+++ test/integration/krane_deploy_test.rb
@@ -1350,7 +1350,7 @@ class KraneDeployTest < Krane::IntegrationTest
       f["missing_volumes.yml"]["Deployment"].first["spec"]["progressDeadlineSeconds"] = 30
       f["cannot_run.yml"]["Deployment"].first["spec"]["replicas"] = 1
     end
-    assert_deploy_failure(result)
+    assert_deploy_failure_or_timeout(result)
 
     bad_probe_timeout = "Deployment/bad-probe: TIMED OUT (progress deadline: 5s)"